Commit 79dfdac ("memcg: make oom_lock 0 and 1 based rather than counter") tried to oom lock the hierarchy and roll back upon encountering an already locked memcg.
The code is confused when it comes to detecting a locked memcg, though, so
it would fail and rollback after locking one memcg and encountering an
unlocked second one.

The result is that oom-locking hierarchies fails unconditionally and that
every oom killer invocation simply goes to sleep on the oom waitqueue
forever.  The tasks practically hang forever without anyone intervening,
possibly holding locks that trip up unrelated tasks, too.

Signed-off-by: Johannes Weiner <jweiner@redhat.com>
Acked-by: Michal Hocko <mhocko@suse.cz>
Acked-by: KAMEZAWA Hiroyuki <kamezawa.hiroyu@jp.fujitsu.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>