ENGR00160936 GPMI : disable the DDR code
We do not support any ONFI or TOGGLE nand now.
So disable the DDR code.

If keep it enabled, the ECC chunk will be 1K for ONFI nand,
this is not compatible with the kobs-ng which uses 512bytes for ECC chunk size.

So disable it now. change it in future.

Signed-off-by: Huang Shijie <b32955@freescale.com>